Rename driver.private -> private.driver
