+++ src/ColorPicker/ColorPicker.private.driver.js
