+++ src/Proportion/Proportion.e2e.js
@@ -5,7 +5,7 @@ import {
 } from 'wix-ui-test-utils/protractor';
 import autoExampleDriver from 'wix-storybook-utils/AutoExampleDriver';
 import { eyesItInstance } from '../../test/utils/eyes-it';
-import { proportionPrivateDriverFactory } from './Proportion.driver.private';
+import { proportionPrivateDriverFactory } from './Proportion.private.driver';
 import { storySettings } from '../../stories/Proportion/storySettings';
 import { PREDEFINED_RATIOS } from './ratios';
