+++ src/RichTextInputArea/RichTextInputArea.private.driver.js
