+++ src/SegmentedToggle/ToggleButton/ToggleButton.spec.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { createUniDriverFactory } from 'wix-ui-test-utils/uni-driver-factory';
 
 import ToggleButton from './ToggleButton';
-import { toggleButtonPrivateDriverFactory } from './ToggleButton.driver.private';
+import { toggleButtonPrivateDriverFactory } from './ToggleButton.private.driver';
 
 describe('ToggleButton', () => {
   const createDriver = createUniDriverFactory(toggleButtonPrivateDriverFactory);
