+++ src/SegmentedToggle/ToggleIcon/ToggleIcon.private.driver.js
